Delete for cluster recreation
Flux must apply the sealed-secrets controller release before any sealed secrets can be interpreted. Because the secret cannot be interpreted, flux chokes and doesn't ever install the sealed-secrets controller. By removing all sealed-secrets flux should be able to install the sealed-secrets controller, and once that is done we can revert the removal of sealed-secrets.

In addition, this secret should not be restored, as it was created with a key which has not been recovered. This has been documented (https://bit.ly/2UQvBJM) as a recovery flaw and mitigated for future cluster restorations